+++ src/eccodes/CMakeLists.txt
@@ -422,10 +422,16 @@ if( eccodes_HAVE_GEOGRAPHY AND eccodes_HAVE_ECKIT_GEO )
         config/LibEccodes.cc
         config/LibEccodes.h
 
+        geo/BasicAngle.cc
+        geo/BasicAngle.h
         geo/EckitMainInit.cc
         geo/EckitMainInit.h
+        geo/Geo.cc
+        geo/Geo.h
         geo/GeoIterator.cc
         geo/GeoIterator.h
+        geo/GribFromSpec.cc
+        geo/GribFromSpec.h
         geo/GribToSpec.cc
         geo/GribToSpec.h )
     list( APPEND ECCODES_EXTRA_LIBRARIES eckit_geo )
