+++ src/eccodes/geo/GribToSpec.cc
@@ -10,7 +10,7 @@
  */
 
 
-#include "geo/GribToSpec.h"
+#include "eccodes/geo/GribToSpec.h"
 
 #include <algorithm>
 #include <cstring>
@@ -38,21 +38,6 @@ const std::vector<double>& gaussian_latitudes(size_t N, bool increasing);
 
 namespace eccodes::geo
 {
-
-
-bool codes_check_error(int e, const char* call)
-{
-    if (e != CODES_SUCCESS) {
-        std::ostringstream os;
-        os << call << ": " << codes_get_error_message(e);
-        throw ::eckit::SeriousBug(os.str());
-    }
-    return true;
-}
-
-
-#define CHECK_ERROR(a, b) codes_check_error(a, b)
-#define CHECK_CALL(a)     codes_check_error(a, #a)
 
 
 namespace
@@ -1159,8 +1144,8 @@ void GribToSpec::json(eckit::JSON& j) const
 }
 
 
-}  // namespace eccodes::geo
-
-
 #undef CHECK_ERROR
 #undef CHECK_CALL
+
+
+}  // namespace eccodes::geo
