+++ src/eccodes/geo/Mercator.cc
