+++ src/eccodes/geo/ReducedGGClassic.cc
