+++ src/eccodes/geo/ReducedGGOctahedral.cc
