--- src/eccodes/geo/grib_from_spec/repres/Lambert.cc
@@ -1,52 +1,0 @@
-/*
- * (C) Copyright 1996- ECMWF.
- *
- * This software is licensed under the terms of the Apache Licence Version 2.0
- * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
- *
- * In applying this licence, ECMWF does not waive the privileges and immunities
- * granted to it by virtue of its status as an intergovernmental organisation nor
- * does it submit to any jurisdiction.
- */
-
-
-namespace eccodes::geo
-{
-
-
-void Lambert_fillGrib(grib_info& info) const
-{
-    info.grid.grid_type = CODES_UTIL_GRID_SPEC_LAMBERT_CONFORMAL;
-
-    Point2 first     = { firstPointBottomLeft() ? x().min() : x().front(), firstPointBottomLeft() ? y().min() : y().front() };
-    Point2 firstLL   = grid().projection().lonlat(first);
-    Point2 reference = grid().projection().lonlat({ 0., 0. });
-
-    info.grid.latitudeOfFirstGridPointInDegrees = firstLL[LLCOORDS::LAT];
-    info.grid.longitudeOfFirstGridPointInDegrees =
-        writeLonPositive_ ? util::normalise_longitude(firstLL[LLCOORDS::LON], 0) : firstLL[LLCOORDS::LON];
-
-    info.grid.Ni = static_cast<long>(x().size());
-    info.grid.Nj = static_cast<long>(y().size());
-
-    info.grid.latitudeOfSouthernPoleInDegrees  = latitudeOfSouthernPoleInDegrees_;
-    info.grid.longitudeOfSouthernPoleInDegrees = longitudeOfSouthernPoleInDegrees_;
-    info.grid.uvRelativeToGrid                 = uvRelativeToGrid_ ? 1 : 0;
-
-    info.extra_set("DxInMetres", std::abs(x().step()));
-    info.extra_set("DyInMetres", std::abs(y().step()));
-    info.extra_set("Latin1InDegrees", reference[LLCOORDS::LAT]);
-    info.extra_set("Latin2InDegrees", reference[LLCOORDS::LAT]);
-    info.extra_set("LoVInDegrees", writeLonPositive_ ? util::normalise_longitude(reference[LLCOORDS::LON], 0)
-                                                     : reference[LLCOORDS::LON]);
-
-    if (writeLaDInDegrees_) {
-        info.extra_set("LaDInDegrees", reference[LLCOORDS::LAT]);
-    }
-
-    // some extra keys are edition-specific, so parent call is here
-    RegularGrid_fillGrib(info);
-}
-
-
-}  // namespace eccodes::geo
