--- src/eccodes/geo/grib_from_spec/repres/ReducedGGOctahedralRotated.cc
@@ -1,25 +1,0 @@
-/*
- * (C) Copyright 1996- ECMWF.
- *
- * This software is licensed under the terms of the Apache Licence Version 2.0
- * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
- *
- * In applying this licence, ECMWF does not waive the privileges and immunities
- * granted to it by virtue of its status as an intergovernmental organisation nor
- * does it submit to any jurisdiction.
- */
-
-
-namespace eccodes::geo
-{
-
-
-void RotatedOctahedral_fillGrib(grib_info& info) const
-{
-    Octahedral_fillGrib(info);
-    rotation_.fillGrib(info);
-    info.grid.grid_type = CODES_UTIL_GRID_SPEC_REDUCED_ROTATED_GG;
-}
-
-
-}  // namespace eccodes::geo
